refactor: error 페이지 구분 및 적용
