+++ frontend/src/App.tsx
@@ -2,7 +2,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { RouterProvider } from 'react-router-dom';
 import { RecoilRoot } from 'recoil';
 import { ThemeProvider } from 'styled-components';
-import Error from 'pages/Error';
+import NotFound from 'pages/Error/NotFound';
 import ErrorBoundary from 'components/@common/ErrorBoundary';
 import ToastList from 'components/@common/Toast/ToastList';
 import ToastProvider from 'contexts/toastContext';
@@ -19,7 +19,7 @@ const App = () => {
       <RecoilRoot>
         <ThemeProvider theme={theme}>
           <QueryClientProvider client={queryClient}>
-            <ErrorBoundary fallback={<Error />}>
+            <ErrorBoundary fallback={<NotFound />}>
               <ToastProvider>
                 <RouterProvider router={router} />
                 <ToastList />
